Moving tag parsing to it's own file
